convert test to testify
